Update patch set 1
Patch Set 1:

Does UploadWizard accept the same query arguments? (such as wpDestFile). And can it show a the link to deletion logs for that title? (which Common's Special:Upload does and is very useful for advanced editors that spider red file links.

Red file links should be rare in articles, so not really a new user entry point.

Patch-set: 1