Always exclude webVideoTranscode jobs from queue processing
The webVideoTranscode jobs will come from the Timed Media Handler
extension. They will be transcoding video files to some mobile friendly
format and sizes, hence they are long jobs that we do not want to run on
the usual servers nor as part of the main queue.

Since the jobrunner loop while nextJobDB.php --wiki=aawiki, we want to
make sure webVideoTranscode are excluded for ALL projects. Hence why it
is added at the bottom of CommonSettings.php (should prevents being
overriden somehow), and why it is always enabled even if a project does
not have the extension enabled.

Change-Id: I769936b4bb53bfe3289c4040f626a09c68e5c1c6