Update patch set 4
Patch Set 4: Do not submit

Just forget about the bug format in comment, it clutter the 'git blame' history making it hard to find out which revision did the change.

Even if we wanted to do that, it should be made another change. So I would prefer you restore the previous patchset which did not introduce all the unrelated changes :-)

Patch-set: 4