Update patch set 1
Patch Set 1:

Maybe I should make the helper function efRaiseAccountCreationThrottle() to just accept an array of parameters such as :

 array(
  'from' => date,
  'to' => date+x
  'IP' => someadd
  'db' => array( ....)  or  'all'
 );

Then do validation of the array parameter and just discard the entry if anything is wrong in it.  'db' could default to 'all'.



Might even be made a proper extension to change any global setting for a time period.  That would probably be better long term.

Patch-set: 1