Update patch set 10
Patch Set 10: There's a problem with this change, please improve

I think it should take an array of DB names, because that has been requested in the past (e.g. bug 37740).

Also, both of the current exceptions have expired. These should be removed from the change or at least commented out as examples or something.

Patch-set: 10