Update patch set 4
Patch Set 4:

Spotted several issues in the extension (mostly small issues). Submitted Ifaf0e574f736264b3d27d376b67fe800904fdeac (pending review).

Should not be deployed until that is merged and  updated in wmf-branches.

In this patchset I updated the config names in advance, so it depends on it now.

Patch-set: 4