Update patch set 2
Patch Set 2: There's a problem with this change, please improve

(2 inline comments)

There are a couple of typos in the commit message.

I'd probably not have used HTML5, but I guess it's ok.
I strongly recommend however to place links to http://www.wikimedia.org/ and http://wikimediafoundation.org/ in the message.

Patch-set: 2