Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

@Platonides: Its not really using HTML5. There are no new HTML5-specific tags or features used. All of this is fully compatible with how browsers have been for the past 10+ years.

The only notable differences is the short doctype, short charset and lack of redundant element wrappers. But those have only become "new" in the HTML5 standard. Real browsers have supported this way before HTML5 was even on the horizon, which is the main reason they became the standard - because it works already and has for a long time.

Regarding extra links, I don't think a link to wikimediafoundation.org is useful in this case because it doesn't help those who are most likely to end up here (tech people testing domains, or visitors of a project seeing it if something went horribly wrong on the server). In both cases they're looking to find Wikimedia projects. I doubt a link to what the foundation itself is all about is useful in this context.

A link to wikimedia.org is already in the page, the logo is clickable. We could make the textual mention a link as well I suppose.

Patch-set: 2