Update patch set 1
Patch Set 1:

No, the v2 logo is the logo they made after some 3D studies in 2010, improving the puzzle and so on.
Using logos from Commons is the de-facto standard since then; reuploading logos on Commons is an easy process which only needs a local sysop; temporary replacements can also be done via CSS and wikis can ask to point back to local wiki.png if they need to.
All wikis must switch to the new logo per WMF's visual identity guidelines, but not all of them did so because creating the new logos is a tedious process.

Patch-set: 1