Update patch set 2
Patch Set 2: There's a problem with this change, please improve

(Not an ops person, take with grain of salt. probably totally wrong).

My understanding is that initializeSettings is loaded before commonSettings, which is where the extension is included. Thus the defaults from the extension would override the settings suggested here. I believe you need to use wmgSubpageSortkeyByNamespace in initializeSettings (note the g) and then put $wgSubpageSortkeyByNamespace = $wmgSubpageSortkeyByNamespace in CommonSettings underneath the require_once for the extension.

------

As an aside (doesn't matter, and not this change) its kind of odd that the variable name $wmfUseSubpageSortkey is used, when the convention seems to be $wmgUseSubpageSortkey.

Patch-set: 2