Update patch set 2
Patch Set 2:

Should I make the change for enwiki alone?
Generally, I don't think it makes much sense to evaluate a filter to 5% if you're only evaluating it on under 50 posts. That means that, at best, a filter can only flag 2 posts before it is automatically shut down (2 flagged posts on a scale of 40 = disable already), this would make it really easy to beat the point of these abusefilters: send in abuse a couple of times -> filter gets disabled -> abuse gets in

Patch-set: 2