Add ability for switching for eqiad-specific configuration
Read the realm (labs or production) from /etc/wikimedia-realm and the
datacenter (pmtpa or eqiad) from /etc/wikimedia-site, and change the
various switches to switch based on $realm instead of cluster (which was
basically $realm trying to pretend to be $datacenter somehow).

Differing configuration for pmtpa versus eqiad are now just a matter of
switching on $datacenter.

Change-Id: Ie0bc55fb5e4a70739030809e7b061ce54ee79ccf