Update patch set 1
Patch Set 1: Looks good to me, approved

Per conversation on IRC yesterday, Jan says in his testing that $wgMaxShellMemory = 300000 was sufficient in all of his testing for 32-bit machines, but that $wgMaxShellMemory = 350000 was the bare minimum that was needed for 64-bit machines.  We all agreed (Tim, Jan, and I) that the combination of the change Tim made and this should be sufficient to fix things.

Patch-set: 1