Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Look good, but...

Just my 2 cents that instead of specific databases, we should take advantage of $site instead if possible (since all are related to Wikivoyage). If we just testing, should be okay, but when we add the other Wikivoyage wikis, it would be good to see this implemented.

Patch-set: 1