(bug 41467) Create new namespace alias WP: for Norwegian (bokmål) Wikipedia
This change implements the namespace alias "WP" for nowiki as per
consensus. Note however that the sysadmin implementing this change
needs to run namespaceDupes.php on nowiki, as there are existing
pages that start with this namespace.

Note: A previous change was attempted (change I74421385), but was
reverted due to issues with the namespaceDupes.php script. Problems
with that script should be sorted out first before this change can
get deployed.

Change-Id: I285fe7c6096d3f45032016449dbf57fcab51bab2