Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Look good (and a nice idea too).

One note however, don't put >62 characters to the first line of the commit message, as it is quite ugly on Gerrit. Summarize the commit on the first line and more notes into the second paragraph, thanks.

Patch-set: 1