Update patch set 2
Patch Set 2: There's a problem with this change, please improve

(2 inline comments)

Some minor questions.

I think we need to warn ops about this change before deploying this. That might stress the imagescalers a bit as well as swift and the upload cache.

Maybe we will have to schedule the change with ops?

Patch-set: 2