(bug 41757) Enable special:import on Hindi Wikipedia
This change enables importing from multiple Wikipedias, Meta, Commons,
Wikispecies and the English language versions for Wikibooks, Wikiquote,
Wikinews, Wikisource and Wikiversity.

Note to reviewer:
1) Consensus was met to implement the transwiki import function on
the Hindi Wikipedia.
2) There was NO discussion regarding what wikis to add as import
source from (other than specifying multiple Wikipedias). If this
is a cause of worry, please feel free to comment on the bug report
about that, and submit a second patch to fix.

Change-Id: Ibd56e2739b948307720947e9571959c9ffc2a041