Update patch set 1
Patch Set 1:

You want to change the behavior with much bigger impact than to properly fix the single-namespace issue. That's inacceptable approach.

This is not a fix to reported bug, but just evil workaround which will annoy those who want to have such preference because they are used to it.

Patch-set: 1