Update patch set 5
Patch Set 5: Abandoned

After working on bug 41133 and finding out how many other things need the same sort of infrastructure for varying configuration based on realm and datacenter, it seems better to just to both this and that in one multi-module changeset instead of continuing to maintain this and then that which would change half of what this does.

See I58901bfdd77b0ce3f273e9391445e43cc776d23c and I7ef35304c18227d95873ceec4bf05d517501a4ce for the replacement for this changeset.

Patch-set: 5