Update patch set 3
Patch Set 3:

This changeset includes the changes from Ie0bc55fb5e4a70739030809e7b061ce54ee79ccf, which was abandoned because this really undoes a lot of what was done there and expands upon it. The biggest difference is that the realm detection code was moved from here in CommonSettings.php into a new file in mediawiki-multiversion so it can be reused more easily.

Patch-set: 3