Update patch set 1
Patch Set 1: There's a problem with this change, please improve

Regardless of the merits of this change or not, we have a handful of places in Wikibase that use bits of ULS code.

We would need to remove those dependencies or else the entire UI (e.g. editing stuff) can break.

Patch-set: 1