Update patch set 1
Patch Set 1: There's a problem with this change, please improve

I am not sure that this is the right way. All other config variables don't set the protocol explicitly.
And you distract users using https.

I guess something in the PHP code has to be modified, setting the protocol.

Patch-set: 1