Language Template fixup definition for UploadWizard.
Note that I've put these into the individual if-branches on purpose.
This is highly dependent on the actual setup of a wiki. The French
wikipedia, for instance, does have Modèle:Tl as the language Template
for Tagalog. The German wikipedia, OTOH, has "Vorlage:tl" as the
"templatelink" template, and has all language templates at
Vorlage:<iso_code>S, so their languageTemplateFixups map will need to be
filled algorithmically. Just something to consider if and when the UW is
rolled out to other wikis but the current three (commons, test, test2).

Change-Id: Ie1b253ef0067fb87a86a36372557396deb352aa8