Update patch set 4
Patch Set 4: Fails; Do not submit

Setting to the worst possible score so gerrit stops trying to merge this and I can stop being spammed.

With 2.6, I'll configure the retry interval to be much wider. 

In the meantime, can we just not submit changes that have dependencies? We *know* this happens.

Patch-set: 4