Update patch set 1
Patch Set 1:

The specific result hoped for from the experiment is a significant increase in users registering, and little enough extra spam that the cost is worthwhile, right?

Do we track number of people who click "create account" and number of people who successfully submit that form so we will be able to see if we benefit?

Patch-set: 1