Update patch set 1
Patch Set 1:

"Do we track number of people who click "create account" and number of people who successfully submit that form so we will be able to see if we benefit?"

Yes.

To clarify here: we've discussed this both in features and with Chris, and what a couple hour test of CAPTCHA removal is going to tell us is the impact on humans. Bots aren't going to notice a change that quick, so a test this short will not be an accurate measurement of how much spam would be released by removing the account creation CAPTCHA. And if the two hour test does release some spammers who lack the dedication to already break our ridiculously easy to break CAPTCHA system, we are more than happy to take charge of Special:Nuke-ing as necessary.

Patch-set: 1