Update patch set 3
Patch Set 3:

Patrick, "what we gain" seems quite clear, data for further research; now also linked.
Perhaps a useful improvement would be adding a conditional with a pre-defined timespan so that 1) it's 3 hours in a meaningful time, not say the middle of the night in USA; 2) it's impossible to forget it enabled for longer.

Patch-set: 3