Update patch set 3
Patch Set 3:

As far as I'm concerned, this test is meant to study the behavior of good-faith humans trying to create accounts.

I don't think it's enough time to study how botmasters would eventually take advantage if we dropped CAPTCHAs.  There will be some amount of spam in just the two hours, but that spam rate can not be extrapolated forward in time convincingly.

Patch-set: 3