Update patch set 3
Patch Set 3:

> Asher: if you have a chance, could you weigh in? It'd be useful to know what you think.

I think the community is more challenging than any captcha and I doubt that captchas are what dissuade potential new editors, even if some need two or three tries.  I'm glad to see this is only planned for the duration of an E3 deployment window, and that the team takes full responsibility for any cleanup that may be needed.

Keep in mind that, even if the chance of spam bots picking this up are slim, if that did happen, it could be taxing to apache cpu and database resources and appear like a malicious attack. Please make sure to notify the ops team ahead of time and to keep us in the loop.

Patch-set: 3