Prevent MediaWiki maintenance scripts from running as privileged users
Maintenance scripts should generally be run using sudo -u apache, which
is available for all wikidev users. Running a maintenance script as a
privileged user risks compromise of the user account. This risk will
become more severe with the plan to allow passwordless sudo to root for
some administrator accounts

Change-Id: I033acd9132e47b840c76da995d14c92e8376775d