Update patch set 1
Patch Set 1:

As I said on IRC, wmf-config/InitialiseSettings-labs.php is loaded on labs AFTER InitialiseSettings.php. The intent is to let us override production settings.

That also mean that most, if not all, of your addition to InitialiseSettings-labs.php are already enabled on beta:

 $ mwscript eval.php --wiki=enwiki
 > var_dump( $wmgMobileFrontend );
 bool(true)
 $

I must confess that setup is REALLY misleading :/

Patch-set: 1