Update patch set 1
Patch Set 1: Abandoned

After a discussion on IRC, Tpt and me agreed such further requests will be handled by extension code, and not by wgExtraNamespaces / wgProofreadPageNamespaceIds.

I'm documenting on wikitech. the procedure to follow.

On a side issue, we're also looking to solve the dual nature (configuration and l10n) of the namespaces, and are considering to populate wmf-config/proofreadpage.php with an array of actual WMF values, the extension code providing a linguistic base, localisable on TranslateWiki.

This change is superseded by I5f7179f0.

Patch-set: 1