Update patch set 1
Patch Set 1:

For some history on the old value, Tim made this comment:
"I think 100/hour is just a conservative limit set when the throttling feature was introduced. IIRC, it was lower to start with, but someone who was running some sort of notification bot complained almost immediately that they needed a higher limit. So we raised the limit to accomodate their needs. Bots have the "noratelimit" right now, so it shouldn't be a problem to reduce the limit."

Patch-set: 1