Update patch set 2
Patch Set 2: Do not submit

Sorry for not having followed up on that issue. Here is a summary of a discussion we had in a restricted mailing list:

It is not reasonable for us to handle different thumbnail sizes per wiki for the following reasons:
- we keep thumbnails forever currently, the more we have the more disk space it takes
- different sizes lower the cache hit rate which in turns cause...
- ... a CPU cost on the cluster to generate a thumbnail, varying the sizes cause more and more thumbnails generations
- whenever a file is updated, we have to purge each thumbnails ever generated

The consensus among the engineering team is to standardize the thumbnails to a few well known formats and have the scaling being done by the web browser.  We could let the mobile clients scale up small thumbnails (to save bandwidth) and desktop scale down larger thumbnails (to get nicer pictures).

So I am sorry to say it, but that changes will not land on the live cluster.

Patch-set: 2