(bug 44251) hardcode $wgDBuser = 'wikiuser'
The DB user name on the WMF cluster is 'wikiuser'. This isn't set
anywhere in wmf-config, but it just so happens to be the software
default in includes/DefaultSettings.php . We should set this value
explicitly in wmf-config, otherwise the site would break completely if
the software default ever changes.

Change-Id: I5e73918ebeea0fb5d4fb0832d18d8aab390436d3