wgRC2UDPPrefix: Use hostname-".org" instead of lang.site
This fixes inconsistently named IRC recent changes channels
that are caused by assuming 'wiki' databases use 'wikipedia.org'
as hostname (where in fact many have dedicates domains or
use subdomains under wikimedia.org instead of wikipeida.org).

I ran a script to extract the current value for $wgRC2UDPPrefix
from each wiki and insert into InitialiseSettings if it is
different than hostname-.org.

So this commit should have no effective change, it merely changes
the default for new wikis from now on. So that instead of
overriding to fix it, it overrides only to use the legacy
naming.

In the future we can easily migrate from the legacy name to
the proper name (bug 28276) by removing the override from
this array.

Change-Id: I249fa712aed7d8b180ec9508f6804df7c624c77a