Update patch set 2
Patch Set 2:

Also, after they are confirmed and agreed on. Need wide announcements to cvn infrastructure and bot maintainers. At least 1 or 2 months I think is reasonable, given the fact that wmf knows almost nothing about it and is in no way invested in it.

Better safe than sorry.

Patch-set: 2