Deploy Global AbuseFilters to Meta-Wiki, MediaWiki and test.wikipedia
Adding on meta is also needed to allow the user interface to work.

As no global rules are currently defined, this should have no effect
on these wikis. This is only to ensure the performance of the code
is as expected.

This deploys the changes from Id69a9d603f9679f838e8691c651a3e9d8461b422

Change-Id: Id04231e161a780a8edf2846217dceceb77c66131