(bug 40759) Let Proofread Page setup namespaces for fi.wikisource
The ids used will be 250 (page) and 252 (index)

Depends on I332b0a60cab5f392e04f9d37b670bb5803050832 which is merged. That needs to be backported and deployed to 1.21wmf9 before this can ber merged and deployed, or this needs to wait until 1.21wmf10 is deployed on fiwikisource.

Change-Id: I072481d57cb23adb607351486003643b84a8e485