Set $wgEventLoggingDBname on test2wiki to test2wiki
This will have the effect of enabling the Schema: namespace on
test2wiki. Because the CodeEditor extension is enabled on test2wiki and
because EventLogging is configured to use CodeEditor for the Schema NS
if the extension is enabled, this configuration change will also allow
us to test CodeEditor functionality on a production-like wiki.

Change-Id: Icfba05ba4b90c4241d18c33fd35a118554830182