Explicitly add Schema NS to test2
As I note in a comment, this namespace should be registered by a hook in
EventLogging, but it isn't, for reasons I can't quite trace. In the interest of
fixing this, I'm going to resort to the same tactic I saw elsewhere in the
file (see commit f2a963f9): registering this namespace explicitly in the configs.

Change-Id: I8ac1eae2456d845afe989052f188aea20f9f5d38