beta: memcached on the two apaches boxes
The aim is to eliminate the deployment-mc box to save up a few resources
on the labs infrastructure.  I have already have the two apaches box
with memcached sessions though they are not being used.

This patch change the beta list of memcached servers from deployment-mc
to the deployment-apache{32,33} instances.

Change-Id: Ia403f766fe0d4fe7f030c82999617eb4c2a11f6b