Move all favicons to bits
In preparation for utilisation of favicon.php and, perhaps eventually,
merging most of the document roots.

Also made the puzzle piece the default favicon for Wiktionary, instead
of the Wikipedia "W" icon as was previously the case.

The old copies need to remain until the Apache configuration for
favicon.php is deployed.

Change-Id: I6b7278255d549eb12dcc2a3a278d4986c2bc806d