Update patch set 1
Patch Set 1: Code-Review-1

Re commit message:
s/puzzle piece/tile/  (or "scrabble-like tile")
"puzzle piece" actually implies the W in puzzle piece as a part of Wikipedia globe

Re config:
Suggesting wmf.ico for default, not wikipedia.ico, to imply it's a project of WMF. Or the community logo. Or even create some new. But not Wikipedia. It should be generic, not specific.

Patch-set: 1