Increase $wgMaxImageArea to 75 Mpx
This will use about 300 MB on ImageMagick Q8. The memory limit is 400MB
per process, now accurately enforced with cgroups, so 300MB is
comfortably underneath this limit.

Note that MaxClients is set to 18 on the image scalers, and they have
12 GB memory, implying an allocation of 680MB per process. So a limit
of 400MB leaves plenty of headroom. So either MaxClients or
$wgMaxShellMemory should be increased to give a better peak utilisation.
They have 24 cores, so neither memory nor CPU will be fully utilised at
peak with the current settings.

Change-Id: Iab0405fae3b0c98cd9cda5d732ce7fcf0b6890f7