Fix use of deprecated configs for Extension:RSS
If you browse to a page on mediawikiwiki that uses the <rss> tag, you'll find
its content replaced with the following notice:

> Extension:RSS -- Error: The deprecated variable $wgRSSAllowedFeeds has been
> detected. Since RSS version 2.0 this variable has to be replaced by
> $wgRSSUrlWhitelist as described in the manual page Extension:RSS.

This change updates our configuration to set the correct variable. The name is
not the only thing that has changed, though -- setting $wgRSSAllowedFeeds to an
empty array (which is what we configured) would previously allow all feeds,
whereas now they need to be explicitly whitelisted. We'll have to add things
like gerritweb by hand, but it's better than allowing content from anywhere.

Change-Id: I556287a9e5f7993a11130701260a76914b8199d6