Update patch set 1
Patch Set 1:

@Tim: Interesting point you bring up. I guess it depends on how you look at the directory. In the current implementation this directory is merely a collection of wrappers with pointers that, based on incoming hostname and wikiversions (multiversion/MWScript + wmf-config), refer to the real mediawiki files.

I also would not disagree with renaming php-* to mw-* in the future, it would make perfect sense and could be done separately.

Whether live-1.5/ should become "mw/" or "mw-multiversion/" depends on how one perceives its meaning and function.

One can see  "php-{version}/" as being the normal mediawiki and live-1.5/ a wrapper for a special occasion, in which case we'd name it "mw-multiversion".

Or consider live-1.5 to be the normal/canonical location, with "php-{version}/" as an implementation detail (foo + foo-wrapper or foo-internal + foo).

I think the latter is more correct in this case.

Neither is a problem or has a clear benefit, it's all details, but since we're talking about it I think it would more semantically/conceptually correct.

Anything that should be agnostic to how we manage deployment currently considers /live-1.5 to be where mediawiki lives and is used from:

* https://en.wikipedia.org/w/index.php is where MediaWiki is used from
* Apache configurations rewrite /wiki to /w/index.php
* /docroots/wikipedia.org/w points to /live-1.5

Imagine the (unlikely?) event that we'd move away from multiversion and go back to a "simple" install. We'd have to update references in all docroots from /apache/common/mw-multiversion to /apache/common/mw.

Again, it wouldn't be a huge pain and we'd likely keep a symlink. The point is not that "mw-multiversion" is wrong, but is it better?

We should separate concerns by abstracting this fact. I think anything pointing to live-1.5 should be agnostic of how we do deployment and in what stage it is. The fact that the current directory name contains a (highly outdated) version number and that we've kept it despite major changes in our infrastructure I find ironic and actually strengthening to this case.

Patch-set: 1