Completely disable AFTv5 on enwiki
The watchlist-induced queries sometimes result in a full table scan.
While I do not yet have any clue on how to tackle that, this will
temporarily completely disable AFTv5 on enwiki. These queries will
still be performed on dewiki, but the total amount of feedback is only
around 26k right now, so it should not really be a problem right away.

Change-Id: I637ef5947832fd215a58a2bb3e0b287ec31043b9