Update patch set 1
Patch Set 1:

I don't know if this would mean their edits would be treated as logged out.  that would be bad and we should confirm.

also would like ops to review or advise about https://gerrit.wikimedia.org/r/#/c/49069/ and https://bugzilla.wikimedia.org/45005 which has the goal of totally eliminate not having the subdomain.  We want it to always be www.wikidata.org and there be no confusion.

Patch-set: 1