Correct entry-point in E3Experiments extension
For a long time, the main entry point to extensions/E3Experiments was
'Experiments.php'. It was a mistake I made early and then had a hard time
reversing. I am about to fix it and am looking to simplify deployment by
temporarily allowing the extension to be included from either entry point.

Change-Id: I01deec11ba17b2f411bfa25eb78dadf0cba35fc8