Remove checkers.php and various obsolete spam regexes
The only defensible thing in checkers.php is the requirement that the UA
be non-empty, but I think that is generally counterproductive since the
normal response to seeing the error message is to fake a browser UA.
Currently there is an issue with the mobile varnish cache which is
causing the blank UA error messages to be cached.

checkers.php contains various ACLs intended to target the vandalism
which was occurring around 2005-2008, a time when shell users were
actively involved in fighting vandalism. The introduction of AbuseFilter
in 2008 made these configuration hacks largely obsolete.

I left in the check for HTML-like links, since that is still often seen
in spam.

Change-Id: Ia1a8685de668364219dd8b5e947938589a2e39ee