Update patch set 1
Patch Set 1: Code-Review-1

Actually, I don't think there's a default for the wait timeout, it'll just give a notice and use null if you don't specify it. It's partly a psychology question -- is it better to fail and require a retry or is it better to wait? I think it's probably better to wait, as long as you don't incite the user or browser to explicitly retry. Obviously it has to be less than the Squid timeout. Probably 15s, like the parser timeout, is about right.

Patch-set: 1