Update patch set 1
Patch Set 1:

In terms of functionality, this seems fine.

In terms of scalability, I don't think this is a great solution. Other (Wikimedia) wikis often try to emulate the English Wikipedia. I think it would be best to avoid a situation in which individual wikis are requesting footer modifications like this, as the bureaucracy and clutter involved (i.e., a bug in Bugzilla, a Gerrit changeset, and additional code in CommonSettings.php) isn't proportionate to the added functionality.

That said, I can't offer a better solution right now, so this is probably fine. Ideally local wiki administrators would be able to modify the footer links without the intervention of developers and system administrators. Perhaps another day.

Patch-set: 1
Label: Code-Review=0