(bug 46990) Add the 'editor' restriction level on pl.wikipedia
This change adds the 'editor' restriction level on pl.wikipedia,
in between the 'autoconfirmed' and 'sysop' protection levels, so
that pages can only be edited by users belonging to the 'editor'
group.

As far as user rights are concerned, this commit:
* Gives the 'editor' right to the 'editor' group so that they
can actually edit pages protected on that middle level;
* Gives the right to the sysop group, so that when a sysop
removes themselves from the 'editor' group, they can still
protect pages to that middle level and edit them;
** This isn't exactly necessary, as all pl.wikipedia sysops
also belong to the 'editor' group, but just in case...
* Gives the 'editor' right to the 'bot' group;
* Assumes that all bureaucrats also belong to the 'sysop'
group.

Bug: 46990
Change-Id: I8c93544133e2327c554ebb1c2af6466042ed2bc7