Update patch set 3
Patch Set 3:

@Nemo said:
> It *will* be a problem for humans if you don't apply the restriction to bot group only... It's easy to make more than 20 edits per minute manually, please don't sabotage editors.

Really, it's easy? Is that something that actually happens during regular editing? Or is this with the help of some script or other?

If this is really a problem, we could allow higher "bursts" - humans are unlikely to sustain such a high rate of editing (that's why we have bots, after all). It should be possible to limit to 200 edits per 600 seconds instead of 20 edits per 60 seconds, which would allow people to make edits fairly quickly, if they don't do it all the time.

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0