noc: Add missing entries to createTxtFileSymlinks.sh
The apache-config and dblist files were still manually created.
These are now tracked in createTxtFileSymlinks.sh as well to
avoid links from going stale or being forgotten about.

Moved fc-list to not be exclusively in docroot/conf, and created
txt links for the cdb files since they trigger a download in Chrome
(instead of viewed raw).

Change-Id: I246a6e2b71a1b3bfac710666b40e1fc25acbfa05