Update patch set 5
Patch Set 5:

First of all, it's very unlikely that MediaWiki:Contact would be defined, but not MediaWiki:Contact-url. Why would you define one, but not the other. If both are undefined, you get no link at all, which is fine.

Second of all, both are already defined on en.wiki (the only wiki this is for), so I'm not sure why it's an issue.

Why does everyone want to prematurely scale this for every wiki? It's only for en.wiki. If another wiki asks if it, we can deal with it then. In the meantime, we shouldn't waste time and code on building a solution for every wiki. This is basic Agile development philosophy. If the Foundation stuck to that philosophy we could accomplish a lot more in a shorter amount of time and with less refactoring.

Patch-set: 5
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0