Update patch set 5
Patch Set 5:

You're right. Ori changed it to be enabled on test instead of en.wiki, which is probably a good idea. I'll see if I can figure out where MediaWiki:Contact is being defined. It's rather mysterious that it isn't in core and doesn't have an extension prefix.

Patch-set: 5