Update patch set 1
Patch Set 1:

Yes, this is true and, quite frankly, a bit surprising. 

There is no reason that I can think of why we have this set as default for autoconfirmed and not for all users, but that's how it is.

I'm only adding this for en.wikipedia since it's what they requested, and because there is no request to enable it globally.

Patch-set: 1
Reviewer: Odder <650@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0