Update patch set 1
Patch Set 1:

Apparently related to the Wikimedia privacy policy: <https://wikimediafoundation.org/w/index.php?title=Privacy_policy&oldid=000000000000000000000000000000000000000000080023#Cookies>.

---

The sites set a temporary session cookie on a visitor's computer whenever a Project page is visited. Readers who do not intend to log in or edit may deny this cookie; it will be deleted at the end of the browser's session. More cookies may be set when one logs in to maintain logged-in status. If one saves a user name or password in one's browser, that information will be saved for up to 30 days, and this information will be resent to the server on every visit to the same Project. Contributors using a public machine who do not wish to show their username to future users of the machine should clear these cookies after use.

---

Patch-set: 1