Update patch set 1
Patch Set 1:

I personally didn't mind 180 days.

But more to the point: it was a little annoying to track this down due to the lack of a bug/announcement. This is a pretty visible part of the user interface. And then when I did pinpoint the change, it still wasn't clear why the change had been made (useless commit message, no code comment). I can't fault you for doing what the legal team says needs to be done, of course, assuming that's what happened here.

Patch-set: 1