Update patch set 1
Patch Set 1:

I don't want to file a separate bug for this, but please do annotate the configuration variable in CommonSettings.php when you get a chance. Just a short comment explaining that the MediaWiki default has been overridden to account for the current Wikimedia Foundation privacy policy. Leaving this configuration bare feels like leaving a mystery (or a trap...) for future file editors.

Patch-set: 1