Update patch set 1
Patch Set 1:

The wikiquote and wikimania favicons seem to be missing the 16x16px resolutions. Isarra contends that 16x16 is unnessasary because her versions would just be purely scaled -- something the web browser should be able to do on it's own. I don't know if that's a correct assertion or not.

Also -- something I just played around with all of these logos might have a tendency to disappear into the background if not on a contrasting background. Not sure what could be done about that.

Patch-set: 1