Update patch set 1
Patch Set 1:

Well, if it's a normal png that just happens to be optimised, it shouldn't be an issue. Generally I've seen problems with IE and files using less common aspects of a format such as a different compression than the default. 

I can't test, though, because windows is simply too unstable for my computers these days - even vmware can't keep it alive anymore.

Patch-set: 1