(bug 48382) Customise Wikimania wiki logos with years
This patch changes $wgLogo for all Wikimania wikis from the
default (and incorrect in case of colours) .gif files into
customised PNG versions with years which are complaint with
the general logo creation guidelines (ie. the text of the
second line is 60% the size of the first line and 60% black).

Obviously, the file pages on Commons are fully protected
against uploads and moves.

As an additional bonus, I created logos for future Wikimanias
up to the year 2020, so there will be no need for any changes
(if the logo doesn't change, of course). When you run out of
logos in the year 2020, please give me a shout!

Bug: 48382
Change-Id: I5715fd861b77289ea8a0a7a337582975d23437c7