(bug 48379) Create custom logos for new private wikis
This patch changes the default logo currently used on fdcwiki,
iegcomwiki, ombudsmenwiki and transitionteamwiki (which has
a rsvg-related rendering issue mentioned in bug 48236) into
custom versions created under the oversight of Thehelpfulone.

The size of the Wikimedia logo in those new logos has been
decreased to match the version used on foundationwiki, in
order to provide as consistent user experience as possible.

Obviously, the file pages on Commons are fully protected
against uploads and moves.

Bug: 48379
Change-Id: I976a779791c6d2a0c6ce76a1e90987b8c2fed1a1