Update patch set 1
Patch Set 1:

No objections on performing purges but a note on the "should not happen": there are indeed discrepancies on thumb containers, with files mostly missing on the Ceph side rather than on the Swift side. This is <= 1%, with just a few containers going up to ~2%.

I've checked a number of those cases by hand and all of the ones I've seen so far were due to Swift hiccups: objects that appear in the container listing but 404 when you attempt to fetch them.

Patch-set: 1
Reviewer: Faidon Liambotis <192@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0