(bug 48308) Change namespace settings for ukwikisource
This patch changes $wgMetaNamespace for the Ukrainian Wikisource
from the default English name into a localised version. It also
adds two two namespaces ('Author' and 'Author_talk') in their
Ukrainian versions as 102 and 103.

I'm also adding the current name of the NS_PROJECT_TALK namespace
to $wgNamespaceAliases for backwards compatibility in case there
are some links using it. Translations for the 'Page' and 'Index'
namespaces are being taken care of in Id781a0b as part of the
ProofreadPage extension.

Bug: 48308
Change-Id: I0b05d2f9d7d62bd5c8e9038cd80be2ae1dde3139