(bug 47574) Change namespace settings for cewiki
This patch changes $wgMetaNamespace and NS_PROJECT_TALK
for the Chechen Wikipedia per community request expressed
in the bug. For backwards compability, the former names
are kept in $wgNamespaceAliases.

I'm also using this opportunity to change $wgSitename
for this wiki, even though it hasn't been requested.
(It doesn't make too much sense to change $wgMetaNamespace
and leave $wgSitename as it was.)

Bug: 47574
Change-Id: I83a116a3e12ca11b5eb2d4305ce243079dc3ad0b