Update patch set 3
Patch Set 3:

Yes, I believe so. That said, I'm going to see if we can get a backlink report for zero.wikipedia.org and its subdomains. Also, wondering if we have info on whether a backlink to a page with no content would add weight to the canonical URI, or on the other extreme trip SEO spam rules.

Patch-set: 3
Label: Verified=0